+++ app/templates/gulp/gulp/config.js
@@ -1,12 +1,10 @@
-'use strict';
-
-var fs = require( 'fs' );
-var glob = require( 'glob' );
+const fs = require( 'fs' );
+const glob = require( 'glob' );
 
 /**
  * Set up file paths
  */
-var loc = {
+const loc = {
   src:  './src',
   dist: './dist',
   lib:  './node_modules', // eslint-disable-line no-sync, no-inline-comments, max-len
@@ -30,7 +28,8 @@ module.exports = {
     ],
     build: [
       'gulpfile.js',
-      'gulp/**/*.js'
+      'gulp/**/*.js',
+      'config/**/*.js'
     ]
   },
   test: {